pyproject.toml
[project]
name = "gnw-evals"
version = "0.1.0"
description = "Global Nature Watch Evaluation Framework"
readme = "README.md"
requires-python = "~=3.13.0,<3.14.0"
dependencies = [
    "click==8.1.7",
    "httpx==0.28.1",
    "langchain-anthropic==1.3.1",
    "langchain-core==1.2.6",
    "pandas==2.3.3",
    "pydantic==2.12.5",
    "pydantic-settings==2.12.0",
    "python-dotenv==1.2.1",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
]

[dependency-groups]
dev = [
    "ruff==0.14.10",
    "pre-commit==4.5.1",
]


[tool.ruff.lint]
select = [
    "COM",  # flake8-commas
    "D",    # pydocstyle
    "F",    # pyflakes
    "I",    # isort
    "RUF",  # ruff-specific
    "UP",   # pyupgrade
]
ignore = [
    "D100",  # Missing docstring in public module
    "D104",
]


[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "asyncio: marks tests as async (deselect with '-m \"not asyncio\"')",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gnw_evals"]

[tool.uv]
default-groups = ["dev"]
